+++ src/main/java/com/authlete/common/api/AuthleteApiImpl.java
@@ -141,6 +141,7 @@ class AuthleteApiImpl implements AuthleteApi
     private static final String VCI_DEFERRED_ISSUE_API_PATH               = "/api/vci/deferred/issue";
     private static final String ID_TOKEN_REISSUE_API_PATH                 = "/api/idtoken/reissue";
     private static final String NATIVE_SSO_API_PATH                       = "/api/nativesso";
+    private static final String NATIVE_SSO_LOGOUT_API_PATH                = "/api/nativesso/logout";
 
     private final String mBaseUrl;
     private final BasicCredentials mServiceOwnerCredentials;
@@ -1871,12 +1872,25 @@ class AuthleteApiImpl implements AuthleteApi
 
 
     @Override
-    public NativeSsoResponse nativeSso(NativeSsoRequest request, Options options) throws AuthleteApiException
+    public NativeSsoResponse nativeSso(
+            NativeSsoRequest request, Options options) throws AuthleteApiException
     {
         // This API call fails because the /nativesso API
         // is unavailable in Authlete 2.x and older versions.
         return callServicePostApi(
                 NATIVE_SSO_API_PATH, request,
                 NativeSsoResponse.class, options);
+    }
+
+
+    @Override
+    public NativeSsoLogoutResponse nativeSsoLogout(
+            NativeSsoLogoutRequest request, Options options) throws AuthleteApiException
+    {
+        // This API call fails because the /nativesso/logout API
+        // is unavailable in Authlete 2.x and older versions.
+        return callServicePostApi(
+                NATIVE_SSO_LOGOUT_API_PATH, request,
+                NativeSsoLogoutResponse.class, options);
     }
 }